pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retail-data-agent"
version = "0.1.0"
description = "Retail data analysis agent over BigQuery with PII masking and oversight."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "langchain-google-genai>=1.0.0",
  "google-cloud-bigquery>=3.13.0",
  "google-auth>=2.28.0",
  "pandas>=2.0.0",
  "db-dtypes==1.2.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.7.0",
  "tenacity>=9.0.0",
  "faiss-cpu>=1.8.0",
  "sqlalchemy>=2.0.0",
  "tabulate>=0.9.0",
  "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-timeout>=2.3.0",
  "mypy>=1.10.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
